fix(Hephaestus): removed _init_ call from compile
